Split extension functions to separated files.
