+++ sdk/src/main/java/co/omise/android/extensions/ViewExtensions.kt
@@ -1,0 +1,9 @@
+package co.omise.android.extensions
+
+import android.view.View
+
+
+fun View.setOnClickListener(action: () -> Unit) {
+    this.setOnClickListener { action() }
+}
+
